Add dns_maradns.sh
MaraDNS is a lightweight self-hosting DNS server. This patch adds
support for adding records to zone files stored on the server in the
format expected by MaraDNS. Path to the file should be exported in
MARA_ZONE_FILE environment variable. To reload the configuration
automatically, the user must provide pid of duende (the daemonization
tool that ships with MaraDNS) in MARA_DUENDE_PID or path to the pid file
(--pid argument to duende) in MARA_DUENDE_PID_PATH.