Check OCSP revocation during --renew/--issue
When renewing or a replacing an existing certificate, acme.sh will check
(via OCSP) whether the existing certificate has been revoked. If it has,
the next renewal time will be ignored and an immediate renewal attempt
will be made.

If the CA does not publish an AIA OCSP URL in the certificate, or if the
OCSP responder does not answer within 10 seconds, no action will be
taken.