Create localcopy deploy-hook
Deploy-hook to very simply copy files to set directories and then execute whatever reloadcmd the admin needs
afterwards. This can be useful for configurations
where the "multideploy" hook (in development) is used
or when an admin wants ACME.SH to renew certs
but needs to manually configure deployment via an external script (e.g. The deploy-freenas script
for TrueNAS Core/Scale
https://github.com/danb35/deploy-freenas/