Allow renewal of staging certs and honor --staging/--test flags.
It is nice to be able to reliably test cert renewals, such as when a cert is
requested in an application using a different method than it is being renewed
as, such as when a cert is obtained through standalone mode, and then expected
to renew through nginx/apache.

Testing renewals by running the cron command swaps out staging cert requests
for production ones.

Operators may want to test the renewal command and append --force --test
of the form:
`acme.sh --cron --home="/home/letsencrypt" --force --test`

make the --test flag (or --staging) flag on this command honor the request to
continue to use staging servers for renewals when requested.

related to #6368